core 28; PM commission as double percentage

+++ src/main/java/com/selfxdsd/storage/SelfContracts.java
@@ -377,9 +377,7 @@ public final class SelfContracts implements Contracts {
                 rec.getValue(SLF_PMS_XDSD.USERNAME),
                 rec.getValue(SLF_PMS_XDSD.PROVIDER),
                 rec.getValue(SLF_PMS_XDSD.ACCESS_TOKEN),
-                BigDecimal.valueOf(
-                    rec.getValue(SLF_PMS_XDSD.COMMISSION).longValue()
-                ),
+                rec.getValue(SLF_PMS_XDSD.COMMISSION).doubleValue(),
                 this.storage
             ),
             this.storage
